# etcdserver: change the snapshot + compact into sync operation