@serathius @ahrtr 
Per the suggestion in (https://github.com/etcd-io/etcd/pull/18235#issuecomment-2209038945), I have changed the snapshot and compact operations to a synchronous process for simplification.

As a first step, I just removed `s.GoAttach(func() {})`. 

I will add benchmark results once all tests pass.