@clement2026 2024-07-05
The benchmark results are impressive.

Reading performance change: -1.67% ~ 27.31%
Writing performance change: -1.12% ~ 28.64%


![compare_read](https://github.com/etcd-io/etcd/assets/27894831/7efa5b62-48ff-43ff-8e46-7e99a1b40e49)
[main.csv](https://github.com/user-attachments/files/16104466/main.csv)
[patch.csv](https://github.com/user-attachments/files/16104468/patch.csv)

The benchmarks were conducted on a cloud VM with 8 vCPUs and 16 GB of memory using the following script:
```bash
export RATIO_LIST="4/1"
export REPEAT_COUNT=3
export RUN_COUNT=50000
./rw-benchmark.sh
```