@serathius 2024-07-05
That doesn't look very good, making snapshot sync creates up to 30% regression. cc @ahrtr