@wenzel2024 2024-07-05
> That doesn't look very good, making snapshot sync creates up to 30% regression. cc @ahrtr

Actually, my observations show a performance improvement of up to 30%. The `patch.csv` shows higher throughput.  @serathius Could you please recheck?😬