@serathius 2024-07-05
> Actually, my observations show a performance improvement of up to 30%. The patch.csv shows higher throughput. @serathius Could you please recheck?😬

Oh yea, sorry I read it incorrectly. I just have been reading benchmark results which show average request duration, and not throughput so made me thing more is worse.