@wenzel2024 2024-07-05
> Oh yea, sorry I read it incorrectly. I just have been reading benchmark results which show average request duration, and not throughput so made me thing more is worse.

lol. I got it. It happens.🤪

As the graph indicates a performance drop with larger value sizes, I am running the `rw-benchmark.sh` script with larger value sizes to verify this issue.