@ahrtr 2024-07-15
Thanks @clement2026 for the test report. The throughput increase up to 30% is a little weird. Theoretically, the performance should be very close. 

From implementation perspective, the only possible reason for the throughput increase  I can think of could be due to the code snippet below, which won't be executed anymore in this PR. Could you double confirm this to help us have a better understanding? e.g. temporarily remove the code snippet on main branch, and then compare with this PR again.

https://github.com/etcd-io/etcd/blob/b433760ca74be6522cb1f219905d32469286210e/server/etcdserver/server.go#L2432-L2440

The CPU usage is very close, which looks fine.